Imlemented cinematics cameras! Updated DBCStorage systems from TC
Former-commit-id: 2af7e8041bd7511cd54422caf4d5d07a632ced2a

+++ src/server/game/playerbot/strategy/actions/InventoryAction.cpp
@@ -135,7 +135,7 @@ void InventoryAction::TellItems(map<uint32, int> itemMap)
 
     items.sort(compare_items);
 
-    uint32 oldClass = -1;
+    uint32 oldClass = uint32(-1);
     for (list<ItemTemplate const*>::iterator i = items.begin(); i != items.end(); i++)
     {
         ItemTemplate const *proto = *i;
